Added Solution zip file

+++ Übungen/Übung_2/Lösung_Jonas/U2_1_c.c
@@ -5,9 +5,9 @@ signed char get_second_equations_result ();
 
 int main ()
 {
-	printf("%i\n", get_first_equations_result());
+	printf("%i\n", (int)get_first_equations_result());
 
-	printf("%i\n", get_second_equations_result());
+	printf("%i\n", (int)get_second_equations_result());
 
 	return 0;
 }
